fixed link to update form on the button. Made a link for now inside of the button. Need to work on the submssion of the edits now and then the delete button
